StopToResturctureServices
- Need to add DTOs
- Need to reduce services amount

+++ RepetitionCore/Models/BasketItem.cs
@@ -4,7 +4,7 @@ namespace RepetitionCore.Models
 {
     public class BasketItem : BaseEntity
     {
-        public int TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public int TotalAmount { get; set; }
         public int BookId { get; set; }
         public int BasketId { get; set; }
